requirements/devolpment.txt
# development.txt
django-debug-toolbar==1.9.1
-r base.txt
